+++ Sources/RichTextKit/Attributes/RichTextAttribute.swift
@@ -3,7 +3,7 @@
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-05-28.
-//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
@@ -12,3 +12,9 @@ import Foundation
  This typealias represents a rich text dictionary key.
  */
 public typealias RichTextAttribute = NSAttributedString.Key
+
+/**
+ This typealias represents a ``RichTextAttribute`` keyed and
+ `Any` valued dictionary.
+ */
+public typealias RichTextAttributes = [RichTextAttribute: Any]