+++ Sources/RichTextKit/Attributes/RichTextAttributeReader+Style.swift
@@ -11,9 +11,12 @@ import Foundation
 public extension RichTextAttributeReader {
 
     /// Get the text styles at a certain range.
-    func richTextStyles(at range: NSRange) -> [RichTextStyle] {
+    func richTextStyles(
+        at range: NSRange
+    ) -> [RichTextStyle] {
         let attributes = richTextAttributes(at: range)
-        let traits = richTextFont(at: range)?.fontDescriptor.symbolicTraits
+        let font = richTextFont(at: range)
+        let traits = font?.fontDescriptor.symbolicTraits
         var styles = traits?.enabledRichTextStyles ?? []
         if attributes.isStrikethrough { styles.append(.strikethrough) }
         if attributes.isUnderlined { styles.append(.underlined) }