+++ Sources/RichTextKit/Attributes/RichTextAttributeWriter+Style.swift
@@ -10,39 +10,37 @@ import Foundation
 
 public extension RichTextAttributeWriter {
 
-    /**
-     Set a rich text style at a certain range.
-
-     The function uses `safeRange(for:)` to handle incorrect
-     ranges, which is not handled by the native functions.
-
-     > Note: When adding logic to the function, make sure to
-     also adjust `setCurrentRichTextStyle`.
-
-     - Parameters:
-       - style: The style to set.
-       - newValue: The new value to set the attribute to.
-       - range: The range to affect, by default the entire text.
-     */
+    ///  Set a rich text style at a certain range.
     func setRichTextStyle(
         _ style: RichTextStyle,
         to newValue: Bool,
-        at range: NSRange? = nil
+        at range: NSRange
     ) {
         let value = newValue ? 1 : 0
-        let range = safeRange(for: range ?? richTextFullRange)
+        let range = safeRange(for: range)
         switch style {
         case .bold, .italic:
-            let styles = richTextStyles(at: range)
-            guard shouldAddOrRemove(style, newValue, given: styles) else { return }
-            guard let font = richTextFont(at: range) else { return }
-            guard let newFont = newFont(for: font, byToggling: style) else { return }
-            setRichTextFont(newFont, at: range)
+            setRichTextBoldOrItalic(style, to: newValue, at: range)
         case .strikethrough:
             setRichTextAttribute(.strikethroughStyle, to: value, at: range)
         case .underlined:
             setRichTextAttribute(.underlineStyle, to: value, at: range)
         }
+    }
+}
+
+private extension RichTextAttributeWriter {
+    
+    func setRichTextBoldOrItalic(
+        _ style: RichTextStyle,
+        to newValue: Bool,
+        at range: NSRange
+    ) {
+        let styles = richTextStyles(at: range)
+        guard shouldAddOrRemove(style, newValue, given: styles) else { return }
+        guard let font = richTextFont(at: range) else { return }
+        guard let newFont = newFont(for: font, byToggling: style) else { return }
+        setRichTextFont(newFont, at: range)
     }
 }