+++ Sources/RichTextKit/Attributes/RichTextAttributeWriter+Superscript.swift
@@ -14,9 +14,9 @@ public extension RichTextAttributeWriter {
     /// Set the superscript level at a certain range.
     func setRichTextSuperscriptLevel(
         _ level: Int,
-        at range: NSRange? = nil
+        at range: NSRange
     ) {
-        setRichTextAttribute(.superscript, to: level)
+        setRichTextAttribute(.superscript, to: level, at: range)
     }
 }
 #endif