+++ Sources/RichTextKit/_Deprecated/RichTextAttributeWriter+ColorDeprecated.swift
@@ -7,7 +7,7 @@ public extension RichTextAttributeWriter {
         _ color: ColorRepresentable,
         at range: NSRange? = nil
     ) {
-        setRichTextAttribute(.backgroundColor, to: color, at: range)
+        setRichTextAttribute(.backgroundColor, to: color, at: range ?? richTextFullRange)
     }
 
     @available(*, deprecated, renamed: "setRichTextColor(_:to:at:)")
@@ -15,7 +15,7 @@ public extension RichTextAttributeWriter {
         _ color: ColorRepresentable,
         at range: NSRange? = nil
     ) {
-        setRichTextAttribute(.foregroundColor, to: color, at: range)
+        setRichTextAttribute(.foregroundColor, to: color, at: range ?? richTextFullRange)
     }
 
     @available(*, deprecated, renamed: "setRichTextColor(_:to:at:)")
@@ -23,7 +23,7 @@ public extension RichTextAttributeWriter {
         _ color: ColorRepresentable,
         at range: NSRange? = nil
     ) {
-        setRichTextAttribute(.strikethroughColor, to: color, at: range)
+        setRichTextAttribute(.strikethroughColor, to: color, at: range ?? richTextFullRange)
     }
 
     @available(*, deprecated, renamed: "setRichTextColor(_:to:at:)")
@@ -31,7 +31,7 @@ public extension RichTextAttributeWriter {
         _ color: ColorRepresentable,
         at range: NSRange? = nil
     ) {
-        setRichTextAttribute(.strokeColor, to: color, at: range)
+        setRichTextAttribute(.strokeColor, to: color, at: range ?? richTextFullRange)
     }
 
     @available(*, deprecated, renamed: "setRichTextColor(_:to:at:)")
@@ -39,6 +39,6 @@ public extension RichTextAttributeWriter {
         _ color: ColorRepresentable,
         at range: NSRange? = nil
     ) {
-        setRichTextAttribute(.underlineColor, to: color, at: range)
+        setRichTextAttribute(.underlineColor, to: color, at: range ?? richTextFullRange)
     }
 }